crawlingAPI/requirements.txt
asgiref==3.2.7
Django==3.0.4
pytz==2019.3
sqlparse==0.3.1
selenium==3.14.1
nltk
By
python-time
Pillow
djangorestframework
django-cors-headers
mysqlclient
